Ignore old move/resize events
This fixes a bug during startup when we learn about a lot of new windows
at once. We resize some of them before learning about others and then
send more resize requests. This is fine, except the resize events from
the first request sometimes arrive after the second request was issued,
causing us to update the model as if the user resized the window to the
size we issued in the first request.

To prevent this we stamp all incoming events with an ID representing the
last request that was seen and forwarded to the app by the app thread.
Events with an outdated stamp are ignored by the reactor.

Also move to a HashMap for doing WindowId lookups in the app thread. We
still have to do a linear scan to look up by window element.